Refactor: 라우트 적용을 위한 페이지 컴포넌트 구조 개선
- Home페이지를 삭제하며 네비게이터는 App컴포넌트로 이동
- ToWatch 페이지 컴포넌트를 새로 생성, root경로에 매칭하여 사용 할 계획

Changes to be committed:
modified:   src/ts/app/App.ts
modified:   src/ts/components/index.ts
deleted:    src/ts/pages/Home.ts
new file:   src/ts/pages/Like.ts
renamed:    src/ts/components/Main.ts -> src/ts/pages/ToWatch.ts
new file:   src/ts/pages/Watched.ts
modified:   src/ts/pages/index.ts
