+++ src/ts/components/index.ts
@@ -1,6 +1,5 @@
 export { default as Modal } from './Modal';
 export { default as Header } from './Header';
-export { default as Main } from './Main';
 export { default as SaveVideoSection } from './SaveVideoSection';
 export { default as Video } from './Video';
 export { default as Button } from './Button';
