+++ src/ts/pages/ToWatch.ts
@@ -1,16 +1,14 @@
-import { EmptyMessage } from '.';
+import { EmptyMessage, SaveVideoSection, Video } from '../components';
 import { createNode } from '../domHelper';
 import store from '../store';
-import { Component } from '../types';
-import SaveVideoSection from './SaveVideoSection';
-import Video from './Video';
+import { CommonProps, Component } from '../types';
 
-interface Props {}
+interface Props extends CommonProps {}
 
-const Main: Component<Props> = ({}) => {
+const ToWatch: Component<Props> = () => {
   const { saveVideoList } = store.getState();
 
-  const $main = createNode('<main class="mt-10"></main>', [
+  const $toWatch = createNode('<main class="mt-10"></main>', [
     SaveVideoSection({
       children: saveVideoList.length
         ? saveVideoList.map((video, index) =>
@@ -29,7 +27,7 @@ const Main: Component<Props> = ({}) => {
     }),
   ]);
 
-  return $main;
+  return $toWatch;
 };
 
-export default Main;
+export default ToWatch;
