+++ src/ts/pages/index.ts
@@ -1,1 +1,1 @@
-export { default as Home } from './Home';
+export { default as ToWatch } from './ToWatch';
